+++ vulture/whitelists/string_whitelist.py
@@ -1,0 +1,10 @@
+import string
+
+string.Formatter.check_unused_args
+string.Formatter.convert_field
+string.Formatter.format
+string.Formatter.format_field
+string.Formatter.get_field
+string.Formatter.get_value
+string.Formatter.parse
+string.Formatter.vformat